generate 구 현

+++ user-service/src/main/java/com/gemini/userservice/dto/GeneratePoseDto.java
@@ -19,5 +19,5 @@ public class GeneratePoseDto {
 
     private List<Long> gemini_seed;
 
-    private Long pose_id;
+    private Integer pose_id;
 }
